factorial is updated for storing Large value

+++ factorialOfNumber.cpp
@@ -1,8 +1,8 @@
 #include<iostream>
 using namespace std;
 
-int factorial(int n){
-    int ans=1;
+long long int factorial(int n){
+    long long int ans=1;
     for(int i=1;i<=n;i++){
         ans = ans*i;
     }
@@ -14,7 +14,7 @@ cout<<"Enter a Number for factorial: ";
  int n;
  cin>>n;
 
- int answer = factorial(n);
+ long long int answer = factorial(n);
 
  cout<<"Factorial of "<<n<<" is: " <<answer;
